third_party/WebKit/LayoutTests/fast/forms/radio/radio-arrow-with-modifier-keys.html

<!DOCTYPE html>                                                                 
                                        
<script src="../resources/common.js"></script>                                  
                                        
<script src="../../../resources/js-test.js"></script>  

<form>
<input type="radio" name="fruit" id="apple">
<input type="radio" name="fruit" id="banana">
<input type="radio" name="fruit" id="cherry">
</form>

<script>
description('Tests behavior of arrow with modifer keys on radio button.');

clickElement(document.getElementById('banana'));
shouldBeTrue('document.getElementById("banana").checked');

eventSender.keyDown('ArrowRight', ['ctrlKey']);
shouldBeTrue('document.getElementById("banana").checked');

eventSender.keyDown('ArrowLeft', ['shiftKey']);
shouldBeTrue('document.getElementById("apple").checked');

eventSender.keyDown('ArrowLeft', ['altKey']);
shouldBeTrue('document.getElementById("apple").checked');
</script>
