third_party/blink/web_tests/platform/generic/fast/forms/radio/radio-arrow-with-modifier-keys-expected.txt

Tests behavior of arrow with modifer keys on radio button.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("banana").checked is true
PASS document.getElementById("banana").checked is true
PASS document.getElementById("apple").checked is true
PASS document.getElementById("apple").checked is true
PASS successfullyParsed is true

TEST COMPLETE

  
